fixing up new errors after package creation
